requirements.txt
# Kafka
confluent-kafka==2.3.0
kafka-python-ng==2.2.2

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# OpenTelemetry - Tracing
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Metrics
prometheus-client==0.19.0
opentelemetry-exporter-prometheus==0.43b0

# Database
sqlalchemy==1.4.48
psycopg2-binary==2.9.9
alembic==1.13.1

# AWS/MinIO
boto3==1.34.25
botocore==1.34.25

# Airflow (for local DAG development)
apache-airflow==2.8.1
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-amazon==8.13.0

# Data Processing
pandas==2.0.3
pydantic==2.5.3
python-dotenv==1.0.0

# Utilities
faker==22.2.0
structlog==24.1.0
httpx==0.26.0
networkx==3.2.1
jinja2==3.1.3
orjson==3.9.12

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3